Swap untouched package names left for compatibility to reflect more accurately what they do + small ddg fix
