+++ plugins/markov/middleware.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/depado/go-b0tsec/configuration"
 	"github.com/depado/go-b0tsec/database"
-	"github.com/depado/go-b0tsec/pluginsinit"
+	"github.com/depado/go-b0tsec/plugins"
 	"github.com/depado/go-b0tsec/utils"
 	"github.com/thoj/go-ircevent"
 )
@@ -19,7 +19,7 @@ const (
 type Middleware struct{}
 
 func init() {
-	m := pluginsinit.Middlewares
+	m := plugins.Middlewares
 	if utils.StringInSlice(middlewareName, configuration.Config.Middlewares) {
 		MainChain = NewChain("main")
 		if err := database.BotStorage.CreateBucket(bucketName); err != nil {
