+++ plugins/title/middleware.go
@@ -13,7 +13,7 @@ import (
 	"golang.org/x/net/html"
 
 	"github.com/depado/go-b0tsec/configuration"
-	"github.com/depado/go-b0tsec/pluginsinit"
+	"github.com/depado/go-b0tsec/plugins"
 	"github.com/depado/go-b0tsec/utils"
 
 	"github.com/rakyll/magicmime"
@@ -30,7 +30,7 @@ var re = regexp.MustCompile(`(?:https?://)(?:www.)?([^/]*).*`)
 type Middleware struct{}
 
 func init() {
-	m := pluginsinit.Middlewares
+	m := plugins.Middlewares
 	if utils.StringInSlice(middlewareName, configuration.Config.Middlewares) {
 		m = append(m, new(Middleware).Get)
 	}
