+++ plugins/urban/plugin.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/depado/go-b0tsec/configuration"
-	"github.com/depado/go-b0tsec/pluginsinit"
+	"github.com/depado/go-b0tsec/plugins"
 	"github.com/depado/go-b0tsec/utils"
 	"github.com/thoj/go-ircevent"
 )
@@ -43,7 +43,7 @@ type Plugin struct {
 
 func init() {
 	if utils.StringInSlice(pluginCommand, configuration.Config.Plugins) {
-		pluginsinit.Plugins[pluginCommand] = new(Plugin)
+		plugins.Plugins[pluginCommand] = new(Plugin)
 	}
 }
