+++ plugins/usercommand/plugin.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/depado/go-b0tsec/configuration"
-	"github.com/depado/go-b0tsec/pluginsinit"
+	"github.com/depado/go-b0tsec/plugins"
 	"github.com/depado/go-b0tsec/utils"
 	"github.com/thoj/go-ircevent"
 )
@@ -17,7 +17,7 @@ const (
 func init() {
 	if utils.StringInSlice(pluginCommand, configuration.Config.Plugins) {
 		CreateBucket()
-		pluginsinit.Plugins[pluginCommand] = new(Plugin)
+		plugins.Plugins[pluginCommand] = new(Plugin)
 	}
 }
