+++ plugins/youtube/middleware.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/depado/go-b0tsec/configuration"
-	"github.com/depado/go-b0tsec/pluginsinit"
+	"github.com/depado/go-b0tsec/plugins"
 	"github.com/depado/go-b0tsec/utils"
 	"github.com/thoj/go-ircevent"
 
@@ -22,7 +22,7 @@ const (
 var ytre = regexp.MustCompile(`(?:https?://)?(?:(?:www\.)?youtube\.com/watch\?.*v=|youtu\.be/)([^&?]{11})`)
 
 func init() {
-	m := pluginsinit.Middlewares
+	m := plugins.Middlewares
 	if utils.StringInSlice(middlewareName, configuration.Config.Middlewares) {
 		m = append(m, new(Middleware).Get)
 	}
